+++ project/src/main/java/lu/uni/adtool/tools/Debug.java
@@ -6,6 +6,9 @@ public final class Debug {
       StackTraceElement[] stack = Thread.currentThread().getStackTrace();
       if (stack.length > 2) {
         msg = stack[2].getFileName() + ":"+ stack[2].getLineNumber() + " " +stack[2].getMethodName() + "(): \t" + msg;
+//         if (stack.length > 3) {
+//         msg += "\n    "+stack[3].getFileName() + ":"+ stack[3].getLineNumber() + " " +stack[3].getMethodName() + "(): \t";
+//         }
       }
       System.out.println(msg);
     }