+++ project/src/main/java/lu/uni/adtool/ui/canvas/SandCanvasHandler.java
@@ -416,7 +416,6 @@ public class SandCanvasHandler extends AbstractCanvasHandler {
                                                               // "").replaceAll("^
                                                               // +| +$| +\n|(
                                                               // )+","$1");
-      menuNode.setName(s);
       ((SandTreeCanvas<?>) canvas).setLabel(menuNode, s);
       ((SandTreeCanvas<?>) canvas).setFocus(menuNode);
     }