+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the ADTool will be documented in this file.
 ## Unreleased
   - added command line examples
+  - fixed regression bug when exporting ranking from command line
+  - renaming node now perserves the node value in domains
 ## [2.0.2] - 2016-03-21 ##
   - added copy/paste of values in Domain View
   - fixed not updating node size after checking "Computed Values" checkbox